Update Kodein to Allow for Different Configs
Allows for debug and release builds to provide custom modules.
Currently this is being used to setup logging on debug builds
but will probably be expanded apon later, like when adding firebase.

Also, to be able to test this on release builds a key was needed.
So the debug key was added to the project (this will allow all
debug builds to use the same signature an not requiring uninstalls
if the signature changes). The release builds will default to the
debug key but can be changed to use the release key and it's values
using the `signing-config` plugin.

Lastly while I was setting this all up I turned on optomization for
release builds.